# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PRS API",
        "description": "API documentation for the PRS application.",
        "contact": {
            "email": "support@prs.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/retailer-orders": {
            "get": {
                "tags": [
                    "Retailer Orders"
                ],
                "summary": "Get all orders for the authenticated retailer",
                "operationId": "5075d7ee92fb32884388ffd987a770c5",
                "responses": {
                    "200": {
                        "description": "List of retailer orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "order_code": {
                                                "type": "string",
                                                "example": "RO-X9Y8Z7"
                                            },
                                            "total_amount": {
                                                "type": "string",
                                                "example": "1500.00"
                                            },
                                            "total_items": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "total_quantity": {
                                                "type": "integer",
                                                "example": 12
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "pending"
                                            },
                                            "notes": {
                                                "type": "string",
                                                "example": "Urgent delivery"
                                            },
                                            "placed_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-10-25 10:00:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized - User is not a retailer"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Retailer Orders"
                ],
                "summary": "Place a new retailer order",
                "operationId": "403e0546298da0f2f242555703901986",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order placed successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/retailer-orders/{id}/update-status": {
            "post": {
                "tags": [
                    "Retailer Orders"
                ],
                "summary": "Update order status (e.g. confirm delivery or cancel)",
                "operationId": "b25309fe7a37c2b804685a62f6325fcb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "delivered",
                                            "cancelled"
                                        ]
                                    },
                                    "cancellation_reason": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully"
                    },
                    "400": {
                        "description": "Invalid status or order state"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}